--- src/sql-parser/src/dialect/mod.rs
@@ -1,49 +1,0 @@
-// Copyright 2018 sqlparser-rs contributors. All rights reserved.
-// Copyright 2019 Materialize, Inc. All rights reserved.
-//
-// This file is derived from the sqlparser-rs project, available at
-// https://github.com/andygrove/sqlparser-rs. It was incorporated
-// directly into Materialize on December 21, 2019.
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License in the LICENSE file at the
-// root of this repository, or online at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-mod ansi;
-mod generic;
-pub mod keywords;
-mod mssql;
-mod mysql;
-mod postgresql;
-
-use std::fmt::Debug;
-
-pub use self::ansi::AnsiDialect;
-pub use self::generic::GenericDialect;
-pub use self::mssql::MsSqlDialect;
-pub use self::mysql::MySqlDialect;
-pub use self::postgresql::PostgreSqlDialect;
-
-pub trait Dialect: Debug {
-    /// Determine if a character starts a quoted identifier. The default
-    /// implementation, accepting "double quoted" ids is both ANSI-compliant
-    /// and appropriate for most dialects (with the notable exception of
-    /// MySQL, MS SQL, and sqlite). You can accept one of characters listed
-    /// in `Word::matching_end_quote` here
-    fn is_delimited_identifier_start(&self, ch: char) -> bool {
-        ch == '"'
-    }
-    /// Determine if a character is a valid start character for an unquoted identifier
-    fn is_identifier_start(&self, ch: char) -> bool;
-    /// Determine if a character is a valid unquoted identifier character
-    fn is_identifier_part(&self, ch: char) -> bool;
-}
