--- src/sql-parser/src/dialect/mysql.rs
@@ -1,41 +1,0 @@
-// Copyright 2018 sqlparser-rs contributors. All rights reserved.
-// Copyright 2019 Materialize, Inc. All rights reserved.
-//
-// This file is derived from the sqlparser-rs project, available at
-// https://github.com/andygrove/sqlparser-rs. It was incorporated
-// directly into Materialize on December 21, 2019.
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License in the LICENSE file at the
-// root of this repository, or online at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-use crate::dialect::Dialect;
-
-#[derive(Debug)]
-pub struct MySqlDialect {}
-
-impl Dialect for MySqlDialect {
-    fn is_identifier_start(&self, ch: char) -> bool {
-        // See https://dev.mysql.com/doc/refman/8.0/en/identifiers.html.
-        // We don't yet support identifiers beginning with numbers, as that
-        // makes it hard to distinguish numeric literals.
-        (ch >= 'a' && ch <= 'z')
-            || (ch >= 'A' && ch <= 'Z')
-            || ch == '_'
-            || ch == '$'
-            || (ch >= '\u{0080}' && ch <= '\u{ffff}')
-    }
-
-    fn is_identifier_part(&self, ch: char) -> bool {
-        self.is_identifier_start(ch) || (ch >= '0' && ch <= '9')
-    }
-}
